final/requirements.txt
streamlit==1.24.1
face-recognition==1.3.0
matplotlib.pyplot==3.7.2
numpy==1.25.0